reorganization of files and types in prep for bringing connections together as a common protocol subsystem
