+++ MeetingNotes/Views/SyncConnectionView.swift
@@ -4,7 +4,7 @@ import SwiftUI
 
 /// A view that displays a sync connection and its state.
 struct SyncConnectionView: View {
-    @ObservedObject var syncConnection: SyncConnection
+    @ObservedObject var syncConnection: BonjourSyncConnection
 
     func stateRepresentationView() -> some View {
         switch syncConnection.connectionState {
