+++ Packages/automerge-repo/Sources/AutomergeRepo/Sync/PeerNetworking/BonjourSyncConnection.swift
@@ -23,7 +23,7 @@ import OSLog
 /// As soon as it is established, it attempts to commence a sync operation (send and expect to receive sync messages).
 /// In addition, it includes an optional `trigger` in its initializer that, when it receives any signal value, kicks off
 /// another attempt to sync the relevant Automerge document.
-public final class SyncConnection: ObservableObject {
+public final class BonjourSyncConnection: ObservableObject {
     /// A unique identifier to track the connections for comparison against existing connections.
     var connectionId = UUID()
     public var shortId: String {
@@ -377,4 +377,4 @@ public final class SyncConnection: ObservableObject {
     }
 }
 
-extension SyncConnection: Identifiable {}
+extension BonjourSyncConnection: Identifiable {}
